feat: implement ESM modules instead of CommonJS and separate code
