+++ constants/constants.js
@@ -10,3 +10,6 @@ export const TABLE_FOOTER =
   "└─────────┴──────────────────────────────────────────────────┴──────────┘";
 
 export const TABLE_HEADER_SKELET = [TABLE_HEADER, TABLE_MAIN, TABLE_DIVIDER];
+
+export const BASIC_COMMANDS = ["cat", "add", "rn", "cp", "mv", "rm"];
+export const NAVIGATION_COMMANDS = [".exit", "up", "cd", "ls"];
